# Simplify and speed up the generation of random base58 strings