@bdewater 2016-07-07
😞 I read over this little bit in the Array#sample docs: "The elements are chosen by using random and unique indices into the array in order to ensure that an element doesn’t repeat itself unless the array already contained duplicate elements." meaning that array-sample is not the right thing to use. I've replaced the PR with the modulo-58 variant.