@matthewd 2016-07-08
```ruby
require 'securerandom'
BASE58_ALPHABET = ('0'..'9').to_a + ('A'..'Z').to_a + ('a'..'z').to_a - ['0', 'O', 'I', 'l']
N = 16

def modulo
  SecureRandom.random_bytes(N).unpack("C*").map do |byte|
    idx = byte % 58
    BASE58_ALPHABET[idx]
  end.join
end

counts = Hash.new(0)
100000.times do
  modulo.each_char { |c| counts[c] += 1 }
end
p counts.sort
```

```
[["1", 31338], ["2", 31295], ["3", 31111], ["4", 30914], ["5", 31390], 
["6", 31280], ["7", 31146], ["8", 31465], ["9", 31120], ["A", 31073], 
["B", 31095], ["C", 31208], ["D", 31087], ["E", 31670], ["F", 31140], 
["G", 31046], ["H", 31342], ["J", 30940], ["K", 31638], ["L", 31364], 
["M", 30938], ["N", 31430], ["P", 30966], ["Q", 31358], ["R", 25335], 
["S", 24872], ["T", 25052], ["U", 24979], ["V", 25149], ["W", 25052], 
["X", 25153], ["Y", 24782], ["Z", 25155], ["a", 25183], ["b", 24867], 
["c", 24894], ["d", 25047], ["e", 24832], ["f", 25135], ["g", 24922], 
["h", 24688], ["i", 25239], ["j", 25250], ["k", 25136], ["m", 24859], 
["n", 24874], ["o", 24942], ["p", 24856], ["q", 24867], ["r", 25179], 
["s", 24971], ["t", 25031], ["u", 24970], ["v", 25368], ["w", 24898], 
["x", 24826], ["y", 25102], ["z", 25181]]
```