@bdewater 2016-07-08
@matthewd TIL. The results speak for themselves but do you have a
link/explanation as to why this happens? Thanks!

On Fri, Jul 8, 2016, 20:07 Matthew Draper <notifications@github.com> wrote:

> Closed #25734 <https://github.com/rails/rails/pull/25734>.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/rails/rails/pull/25734#event-717330782>, or mute the
> thread
> <https://github.com/notifications/unsubscribe/AAeS74Xiy2a0-rJiGAIJJfAWYM62QHB9ks5qTpHOgaJpZM4JG4nw>
> .
>